requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv